refactor: 为方便测试还是把潜能选择模块改为识别了
