+++ agent/custom/action/__init__.py
@@ -1,12 +1,10 @@
 from .climb_tower_shop import *
-from .climb_tower_potential import *
 from .climb_tower_preparation import *
 from .climb_tower_loop import *
 from .invite import *
 from .fight import *
 
 __all__ = [
-    "ChoosePotentialAction",
     "ShopAction",
     "EnhanceAction",
     "InviteAuto",
