+++ agent/custom/reco/__init__.py
@@ -1,5 +1,9 @@
 from .climb_tower_quiz import *
+from .climb_tower_potential import *
+from .climb_tower_potential_bag import *
 
 __all__ = [
-    "QuizRecognition"
+    "QuizRecognition",
+    "ChoosePotentialRecognition",
+    "PotentialBagTestRecognition"
 ]
