+++ assets/resource/base/pipeline/climb_tower/climb_tower_agent_sub.json
@@ -829,5 +829,13 @@
       "type": "Swipe"
     },
     "post_wait_freezes": 500
+  },
+  "星塔_测试用节点_agent": {
+    "recognition": {
+      "type": "Custom",
+      "param": {
+        "custom_recognition": "bag_test_recognition"
+      }
+    }
   }
 }
